+++ src/modsim2/attack/__init__.py
@@ -1,2 +1,2 @@
-from modsim2.attack.attack import generate_over_combinations  # noqa: F401
+from modsim2.attack.adversarial_images import generate_over_combinations  # noqa: F401
 from modsim2.attack.transfer import compute_transfer_attack  # noqa: F401
