+++ src/modsim2/attack/adversarial_images.py
